# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LuaBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List all scripts owned by current user
      responses:
        "200":
          description: List of scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Script"
    post:
      operationId: createScript
      tags: [scripts]
      summary: Create a new script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptInput"
      responses:
        "201":
          description: Created script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get a script by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateScript
      tags: [scripts]
      summary: Update a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptUpdate"
      responses:
        "200":
          description: Updated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /scripts/{id}/toggle:
    post:
      operationId: toggleScript
      tags: [scripts]
      summary: Enable or disable a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated script status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
  /panels:
    get:
      operationId: listPanels
      tags: [panels]
      summary: List all panels owned by current user
      responses:
        "200":
          description: List of panels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Panel"
    post:
      operationId: createPanel
      tags: [panels]
      summary: Create a new Discord panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelInput"
      responses:
        "201":
          description: Created panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
  /panels/{id}:
    get:
      operationId: getPanel
      tags: [panels]
      summary: Get a panel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePanel
      tags: [panels]
      summary: Update a panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelUpdate"
      responses:
        "200":
          description: Updated panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
    delete:
      operationId: deletePanel
      tags: [panels]
      summary: Delete a panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /panels/{id}/send:
    post:
      operationId: sendPanel
      tags: [panels]
      summary: Send panel embed to a Discord channel via bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelSendInput"
      responses:
        "200":
          description: Panel sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /servers/{id}:
    delete:
      operationId: disconnectServer
      tags: [servers]
      summary: Disconnect a Discord server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Script:
      type: object
      required: [id, ownerId, name, version, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        version:
          type: string
        status:
          type: string
          enum: [active, disabled]
        content:
          type: ["string", "null"]
        obfuscatedContent:
          type: ["string", "null"]
        createdAt:
          type: string
    ScriptInput:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: string
          minLength: 1
        content:
          type: string
    ScriptUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: string
          minLength: 1
        status:
          type: string
          enum: [active, disabled]
        content:
          type: string
    Panel:
      type: object
      required: [id, ownerId, scriptId, name, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        scriptId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        discordServerId:
          type: ["string", "null"]
        channelId:
          type: ["string", "null"]
        messageId:
          type: ["string", "null"]
        buyerRoleId:
          type: ["string", "null"]
        requiredRoles:
          type: array
          items:
            type: string
        createdAt:
          type: string
    PanelInput:
      type: object
      required: [scriptId, name]
      properties:
        scriptId:
          type: integer
        name:
          type: string
          minLength: 1
        description:
          type: string
        discordServerId:
          type: string
        channelId:
          type: string
        buyerRoleId:
          type: string
        requiredRoles:
          type: array
          items:
            type: string
    PanelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        discordServerId:
          type: string
        channelId:
          type: string
        buyerRoleId:
          type: string
        requiredRoles:
          type: array
          items:
            type: string
    PanelSendInput:
      type: object
      required: [channelId]
      properties:
        channelId:
          type: string
...